Refactor: MemberSeat 패키지 분리
