+++ src/main/java/ticketaka/mtvs3_final_backend/concert/command/domain/model/ConcertStatus.java
@@ -1,6 +1,6 @@
 package ticketaka.mtvs3_final_backend.concert.command.domain.model;
 
 public enum ConcertStatus {
-    // 준비 중, 예약 중, 대기 중, 공연 중
-    PREPARING, RESERVING, WAITING, PERFORMING
+    // 준비 중, 예약 중, 공연 중
+    PREPARING, RESERVING, PERFORMING
 }
