+++ src/main/java/ticketaka/mtvs3_final_backend/memberseat/command/domain/model/MemberSeatStatus.java
@@ -1,4 +1,4 @@
-package ticketaka.mtvs3_final_backend.seat.command.domain.model;
+package ticketaka.mtvs3_final_backend.memberseat.command.domain.model;
 
 public enum MemberSeatStatus {
     // 접수 중, 예약 대기, 예약
