+++ src/main/java/ticketaka/mtvs3_final_backend/seat/command/domain/repository/SeatRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ticketaka.mtvs3_final_backend.concert.command.domain.model.Concert;
-import ticketaka.mtvs3_final_backend.seat.command.domain.model.MemberSeatStatus;
+import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeatStatus;
 import ticketaka.mtvs3_final_backend.seat.command.domain.model.Seat;
 import ticketaka.mtvs3_final_backend.seat.command.domain.model.SeatStatus;
