+++ src/main/java/ticketaka/mtvs3_final_backend/seat/query/dto/SeatInfoResponseDTO.java
@@ -1,6 +1,6 @@
 package ticketaka.mtvs3_final_backend.seat.query.dto;
 
-public class SeatResponseDTO {
+public class SeatInfoResponseDTO {
 
     // 좌석 정보 조회
     public record getSeatInfoDTO(
